+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [2.4.2] - 2026-09-08
+
+### Fixed
+- Fixed `MergedMesh.split_mesh()` regression with uninitialized loop UV data.
+- Fixed `test_mesh_tools` unit tests to reflect new `SplitMeshDef` layout (`kwargs` replaced).
+
 ## [2.4.1] - 2026-09-07
 
 ### Added